+++ Kartographer/CelestialBodyData.cs
@@ -8,6 +8,8 @@ using System.Collections.Generic;
 using UnityEngine;
 using KSP.IO;
 
+using ClickThroughFix;
+
 namespace Kartographer
 {
 	[KSPAddonImproved (KSPAddonImproved.Startup.Flight | KSPAddonImproved.Startup.TrackingStation, false)]
@@ -97,7 +99,7 @@ namespace Kartographer
 		{
 			if (_active && !_hidden) {
 				if (KartographSettings.Instance.UseKspSkin) GUI.skin = HighLogic.Skin;
-				_windowPos = GUILayout.Window (_winID, _windowPos, OnWindow, "Celestial Body Data");
+				_windowPos = ClickThruBlocker.GUILayoutWindow (_winID, _windowPos, OnWindow, "Celestial Body Data");
 				if (_windowPos.Contains (Event.current.mousePosition)) {
 					ControlLock ();
 				} else {