+++ Kartographer/VesselSelect.cs
@@ -7,6 +7,8 @@ using System.Collections.Generic;
 using UnityEngine;
 using KSP.IO;
 
+using ClickThroughFix;
+
 namespace Kartographer
 {
 	[KSPAddonImproved (KSPAddonImproved.Startup.Flight, false)]
@@ -150,7 +152,7 @@ namespace Kartographer
 		{
 			if (_active && !_hidden) {
 				if (KartographSettings.Instance.UseKspSkin) GUI.skin = HighLogic.Skin;
-				_windowPos = GUILayout.Window (_winID, _windowPos, OnWindow, "Vessel Select");
+				_windowPos = ClickThruBlocker.GUILayoutWindow (_winID, _windowPos, OnWindow, "Vessel Select");
 				if (_windowPos.Contains (Event.current.mousePosition)) {
 					ControlLock ();
 				} else {